refactor!: separate adapter runtime/kineokit concerns
in the future, you will need to pass an adapter to kineokit in its config.
