+++ packages/kineo/tests/adapter/neo4j.test.ts
@@ -1,7 +1,12 @@
 import { describe, test, expect, vi, beforeAll, afterAll } from "vitest";
 import neo4j from "neo4j-driver";
-import { Neo4jAdapter, auth, toNative, collectEdges } from "@/adapters/neo4j";
-import { defineSchema, field, model } from "@/schema";
+import {
+  neo4jAdapter,
+  type Neo4jAdapter,
+  auth,
+  toNative,
+  collectEdges,
+} from "@/adapters/neo4j";
 
 // Neo4j connection settings
 const NEO4J_URL = "bolt://localhost:7687";
@@ -53,11 +58,11 @@ describe("auth()", () => {
   });
 });
 
-describe("Neo4jAdapter (integration)", () => {
+describe("neo4jAdapter (integration)", () => {
   let adapter: Neo4jAdapter;
 
   beforeAll(async () => {
-    adapter = Neo4jAdapter({
+    adapter = neo4jAdapter({
       url: NEO4J_URL,
       auth: {
         type: "basic",
@@ -111,90 +116,6 @@ describe("Neo4jAdapter (integration)", () => {
     const edge = result.edges![0];
     expect(edge.type).toBe("KNOWS");
     expect(edge.props.since).toBe(2020);
-  });
-
-  // ---------------------------------------------------------------------------
-  // pull()
-  // ---------------------------------------------------------------------------
-
-  test("pull() extracts labels, properties, and relationships", async () => {
-    await adapter.session.run(`
-      CREATE (u:User {name: "Alice", age: 25})
-      CREATE (p:Post {title: "Hello"})
-      CREATE (u)-[:WROTE]->(p)
-    `);
-
-    const pulled = await adapter.pull!();
-
-    expect(pulled).toBeDefined();
-    expect(pulled.schema.User).toBeDefined();
-    expect(pulled.schema.Post).toBeDefined();
-
-    // Sampled properties
-    expect(pulled.schema.User.$shape.name).toBeDefined();
-    expect(pulled.schema.User.$shape.age).toBeDefined();
-
-    // Sampled relationships
-    expect(pulled.schema.User.$shape.WROTE).toBeDefined();
-    expect(pulled.schema.Post.$shape.WROTE).toBeDefined();
-  });
-
-  // ---------------------------------------------------------------------------
-  // push()
-  // ---------------------------------------------------------------------------
-
-  test("push() applies constraints & indexes without throwing", async () => {
-    const schema = defineSchema({
-      User: model("User", {
-        id: field.string().id().required(),
-        name: field.string().required(),
-      }),
-    });
-
-    await expect(adapter.push!(schema)).resolves.not.toThrow();
-  });
-
-  // ---------------------------------------------------------------------------
-  // generate()
-  // ---------------------------------------------------------------------------
-
-  test("generate() detects added fields and produces migration entries", async () => {
-    const prev = defineSchema({
-      User: model("User", {}),
-    });
-
-    const cur = defineSchema({
-      User: model("User", {
-        age: field.int().default(0),
-      }),
-    });
-
-    const migrations = await adapter.generate!(prev, cur);
-
-    expect(migrations.length).toBeGreaterThan(0);
-    expect(
-      migrations.some((m) => m.type === "command" && m.command.includes("SET")),
-    ).toBe(true);
-  });
-
-  // ---------------------------------------------------------------------------
-  // deploy() + status()
-  // ---------------------------------------------------------------------------
-
-  test("deploy() stores migration metadata and status() returns completed", async () => {
-    const hash = "test-hash-123";
-    const migration = "CREATE (:MetaTest {x: 1})";
-
-    await adapter.deploy!(migration, hash);
-
-    const state = await adapter.status!(migration, hash);
-
-    expect(state).toBe("completed");
-  });
-
-  test("status() returns pending if migration hash not found", async () => {
-    const state = await adapter.status!("", "non-existing-hash");
-    expect(state).toBe("pending");
   });
 
   // TODO integration tests with client
