+++ packages/kineokit/package.json
@@ -2,10 +2,15 @@
   "name": "kineokit",
   "version": "0.1.0",
   "description": "Migration manager for Kineo.",
-  "main": "./dist/main.mjs",
   "bin": {
     "kineokit": "./dist/main.mjs",
     "kineo": "./dist/main.mjs"
+  },
+  "exports": {
+    "./adapters/neo4j": {
+      "types": "./dist/neo4j.d.mts",
+      "import": "./dist/neo4j.mjs"
+    }
   },
   "scripts": {
     "kit": "jiti src/main.ts",
@@ -23,8 +28,16 @@
   "packageManager": "pnpm@10.28.0",
   "dependencies": {
     "convoker": "^0.3.4",
-    "jiti": "^2.6.1",
-    "kineo": "workspace:*"
+    "jiti": "^2.6.1"
+  },
+  "peerDependencies": {
+    "kineo": "workspace:*",
+    "neo4j-driver": "^6.0.0"
+  },
+  "peerDependenciesMeta": {
+    "neo4j-driver": {
+      "optional": true
+    }
   },
   "devDependencies": {
     "kineo": "workspace:*"
