+++ tsconfig.json
@@ -3,7 +3,6 @@
     "target": "es2020",
     "module": "es2020",
     "moduleResolution": "bundler",
-    "customConditions": ["development"],
     "noEmit": true,
     "strict": true,
     "esModuleInterop": true,
