+++ vitest.config.ts
@@ -2,7 +2,6 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   resolve: {
-    conditions: ["development"],
     alias: {
       "@": "/src",
       "@tests": "/tests",
